完成了SPWM驱动，可以获得真实电角度，完成永磁同步电机

+++ Euler_FOC_SPWM/MDK-ARM/BSP/BSP_TIM.h
@@ -10,7 +10,7 @@
 #include <stdint.h>
 
 
-#include "BSP_LED.h"
+
 
 
 
@@ -20,7 +20,7 @@ void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim);
 
 
 extern volatile uint16_t BSP_time_TIM6;
-extern volatile uint8_t BSP_time_TIM1;
+extern volatile uint16_t BSP_time_TIM1;
 
 #endif
